slim down GrpcReflection module by moving the __using__ macro
