+++ examples/helloworld/lib/reflection_server.ex
@@ -1,5 +1,5 @@
 defmodule Helloworld.Reflection.Server do
-  use GrpcReflection,
+  use GrpcReflection.Server,
     version: :v1,
     services: [Helloworld.Greeter.Service]
 end
